Create Analyze Storage Analytics metrics
Microsoft provides the WindowsAzure.Storage client library that you can use to retrieve the metrics for a storage account. The key class in this library is CloudAnalyticsClient. You can use this class to run queries over the metrics held in the $Metrics tables. The code below shows an example that fetches all the metrics for each service in a storage account.